// src/clients/roman-numeral-client/models/InputValidationError.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Roman Numeral API
 * API for Roman numeral conversions
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { ValidationErrorDetail } from './ValidationErrorDetail';
import {
    ValidationErrorDetailFromJSON,
    ValidationErrorDetailFromJSONTyped,
    ValidationErrorDetailToJSON,
    ValidationErrorDetailToJSONTyped,
} from './ValidationErrorDetail';

/**
 * 
 * @export
 * @interface InputValidationError
 */
export interface InputValidationError {
    /**
     * The error message.
     * @type {string}
     * @memberof InputValidationError
     */
    message: string;
    /**
     * Collection of further details.
     * @type {Array<ValidationErrorDetail>}
     * @memberof InputValidationError
     */
    errorDetails: Array<ValidationErrorDetail>;
}

/**
 * Check if a given object implements the InputValidationError interface.
 */
export function instanceOfInputValidationError(value: object): value is InputValidationError {
    if (!('message' in value) || value['message'] === undefined) return false;
    if (!('errorDetails' in value) || value['errorDetails'] === undefined) return false;
    return true;
}

export function InputValidationErrorFromJSON(json: any): InputValidationError {
    return InputValidationErrorFromJSONTyped(json, false);
}

export function InputValidationErrorFromJSONTyped(json: any, ignoreDiscriminator: boolean): InputValidationError {
    if (json == null) {
        return json;
    }
    return {
        
        'message': json['message'],
        'errorDetails': ((json['errorDetails'] as Array<any>).map(ValidationErrorDetailFromJSON)),
    };
}

export function InputValidationErrorToJSON(json: any): InputValidationError {
    return InputValidationErrorToJSONTyped(json, false);
}

export function InputValidationErrorToJSONTyped(value?: InputValidationError | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'message': value['message'],
        'errorDetails': ((value['errorDetails'] as Array<any>).map(ValidationErrorDetailToJSON)),
    };
}